Restore bigint
Revert 3b8e3dcfc4017d01663b58d2ae8da1892e2fac4c

+++ db_converter.py
@@ -137,7 +137,7 @@ def parse(input_filename, output_filename):
                     type = "integer"
                     set_sequence = True
                 elif type.startswith("bigint("):
-                    type = "integer"
+                    type = "bigint"
                     set_sequence = True
                 elif type == "longtext":
                     type = "text"
